Core, Orc: Remove deprecated partition stats read functionality
