+++ .palantir/revapi.yml
@@ -517,6 +517,62 @@ acceptedBreaks:
     - code: "java.class.removed"
       old: "class org.apache.iceberg.data.PartitionStatsHandler"
       justification: "Removing deprecated code for 1.11.0"
+  "1.11.0":
+    org.apache.iceberg:iceberg-core:
+    - code: "java.class.removed"
+      old: "class org.apache.iceberg.PartitionStats"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.DATA_FILE_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.DATA_RECORD_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.DV_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.EQUALITY_DELETE_FILE_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.EQUALITY_DELETE_RECORD_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.LAST_UPDATED_AT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.LAST_UPDATED_SNAPSHOT_ID"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.POSITION_DELETE_FILE_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.POSITION_DELETE_RECORD_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.SPEC_ID"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.TOTAL_DATA_FILE_SIZE_IN_BYTES"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removed"
+      old: "field org.apache.iceberg.PartitionStatsHandler.TOTAL_RECORD_COUNT"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removedWithConstant"
+      old: "field org.apache.iceberg.PartitionStatsHandler.PARTITION_FIELD_ID"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.field.removedWithConstant"
+      old: "field org.apache.iceberg.PartitionStatsHandler.PARTITION_FIELD_NAME"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.method.removed"
+      old: "method org.apache.iceberg.Schema org.apache.iceberg.PartitionStatsHandler::schema(org.apache.iceberg.types.Types.StructType,\
+        \ int)"
+      justification: "Removed deprecated functionality for partition stats"
+    - code: "java.method.removed"
+      old: "method org.apache.iceberg.io.CloseableIterable<org.apache.iceberg.PartitionStats>\
+        \ org.apache.iceberg.PartitionStatsHandler::readPartitionStatsFile(org.apache.iceberg.Schema,\
+        \ org.apache.iceberg.io.InputFile)"
+      justification: "Removed deprecated functionality for partition stats"
   "1.2.0":
     org.apache.iceberg:iceberg-api:
     - code: "java.field.constantValueChanged"
